refactor: split create workflows by domain
