+++ src/tasks/composer/workspace.ts
@@ -1,0 +1,3 @@
+import type { PrismaWorkspace } from "../../result";
+
+export const getWorkspaceLabel = (workspace: PrismaWorkspace) => workspace.name ?? workspace.id;
